refactoring dir structure
